+++ lib/packages/wikibase/data-model/src/Entity/PropertyId.php
@@ -83,6 +83,8 @@ class PropertyId extends EntityId implements Int32EntityId {
 	public function unserialize( $serialized ) {
 		$array = json_decode( $serialized );
 		$this->serialization = is_array( $array ) ? $array[1] : $serialized;
+		list ( $this->repositoryName, $this->localPart ) =
+			self::extractRepositoryNameAndLocalPart( $this->serialization );
 	}
 
 	/**