+++ lib/packages/wikibase/data-model/tests/fixtures/CustomEntityId.php
@@ -28,6 +28,8 @@ class CustomEntityId extends EntityId {
 	 */
 	public function unserialize( $serialized ) {
 		$this->serialization = $serialized;
+		$this->repositoryName = '';
+		$this->localPart = $serialized;
 	}
 
 	/**